docker/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.0
    # via -r requirements.in
fastapi-cli[standard]==0.0.5
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.12
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.2
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.6
    # via fastapi
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
uvicorn[standard]==0.31.0
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
langchain==0.3.21
langchain-community==0.3.20
elasticsearch==8.17.2
sentence-transformers==3.4.1
spacy==3.8.4
pdfminer-six==20250324
morfeusz2==1.99.9
nltk==3.9.1
beautifulsoup4==4.13.3
openai==1.68.2
accelerate==1.3.0
sentencepiece==0.2.0
protobuf==5.29.1